Added Parameters class
The parameters class reads the entire configuration file
line by line, and cuts each line into its substrings (i.e "nx 5 lattice
length" -> ["nx", "5", "lattice", "length"]).
The first substring is matched to a parameter name (i.e "nx") while the
second is used as the parameter's value (5). The rest of the substrings
are ignored. The parameter is then converted into float. If the value should
be an int, the converted value is then cast into an int.

Once a parameter is matched, the corresponding parameter is set. This parameter
is always named by m_<parametername>. A boolean is set to ensure that the
parameter is set. If not all parameters are set once the file is read, a
runetimeerror is thrown.

+++ Lattice/lattice.cpp
@@ -24,8 +24,6 @@ std::string Lattice::xyzString()
         double phi = node->phi();
         xyz  << r[0] << " " << r[1] << " " << r[2] << " " << cos(phi) << " " << sin(phi) << " " << 0 << "\n";
     }
-
-//    std::cout << xyz.str() << endl;
     return xyz.str();
 }
 
@@ -33,40 +31,26 @@ void Lattice::step(double dt)
 {
     omp_set_num_threads(4);
 #pragma omp flush(dt)
-    //for (auto & node : nodes){
-    //    node->vvstep1(dt);
-    //}
 
 #pragma omp parallel for
-    for (int i = 0; i<nodes.size(); i++)
+    for (size_t i = 0; i<nodes.size(); i++)
     {
-//        double dt2 = dt;
         nodes[i]->vvstep1(dt);
     }
 
     m_t += dt*0.5;
 
-//    for (auto & node : nodes){
-//        node->updateForcesAndMoments();
-//    }
 #pragma omp parallel for
-    for (int i = 0; i<nodes.size(); i++)
+    for (size_t i = 0; i<nodes.size(); i++)
     {
 
         nodes[i]->updateForcesAndMoments();
     }
 
-
-//    for (auto & node : nodes){
-//        node->vvstep2(dt);
-//    }
-
 #pragma omp parallel for
-    for (int i = 0; i<nodes.size(); i++)
+    for (size_t i = 0; i<nodes.size(); i++)
     {
-//        double dt2 = dt;
         nodes[i]->vvstep2(dt);
     }
     m_t += dt*0.5;
 }
-
